[bridge] Simplify block interface
We do not need to have the hash and next_hash as a field on the Block itself.
Those values can be calculated on the go, as they need to be stored as part of
contract data when certain criteria are met. Also this means that constructors
for Block and BlockHeaderInnerLite are no longer needed as they are primarily
used for converting test input (which are in strings) to expected bytes input.

Test plan:
./build.sh
cargo test
./deploy.sh
Verified on the separate branch that the relayer can init_with_block with
the changed contract.

Changes in the relayer (all needed fields sent as bytes) will be sent in a
separate commit