[Dev tools] Enable reseting light client state for dev purposes
When compiled with:
RUSTFLAGS='--cfg reset' cargo build --target wasm32-unknown-unknown --release
The reset_state method is available to the contract owner, it should only be
used when debugging.

Test plan:
Tried deploying with and without the compilation flag.
When deployed with the reset method, verified the contract can be redeployed,
and it will work again from clear state.